semi working names an chat, improved version in the works
