--- backend/persistence.js
@@ -1,2 +1,0 @@
-export const games = {};
-export const disconnectTimers = {};
